Part of database interaction added to User Space. (It's probably full of bugs. I'll start to write the unit test as soon as I'll know what to do with the database.)

+++ src/userspace/src/main/java/cz/filmtit/userspace/Document.java
@@ -37,7 +37,7 @@ public class Document {
     public Document() {
     }
 
-    private int databaseId;
+    private Long databaseId;
     private String movieTitle;
     private int year;
     private List<Chunk> chunks;
@@ -51,11 +51,11 @@ public class Document {
     private MediaSource mediaSource;
     private boolean finished;
 
-    public int getDatabaseId() {
+    public Long getDatabaseId() {
         return databaseId;
     }
 
-    public void setDatabaseId(int databaseId) {
+    public void setDatabaseId(Long databaseId) {
         this.databaseId = databaseId;
     }
